extension: Reorganize capabilities
