+++ crates/extension/src/capabilities.rs
@@ -1,0 +1,16 @@
+mod download_file_capability;
+mod process_exec_capability;
+
+pub use download_file_capability::*;
+pub use process_exec_capability::*;
+
+use serde::{Deserialize, Serialize};
+
+/// A capability for an extension.
+#[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
+#[serde(tag = "kind", rename_all = "snake_case")]
+pub enum ExtensionCapability {
+    #[serde(rename = "process:exec")]
+    ProcessExec(ProcessExecCapability),
+    DownloadFile(DownloadFileCapability),
+}
